EPMLSTRTAI-359: extracted common methods to abstract annotation class
